+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -418,7 +418,7 @@ class CreateDatasourceResponseBody(BaseModel):
     metadata: CreateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY4MCVQVJ1TA3GYJX5078VM9"
+        "01KY4TS7NN0J34NZQHT0ZR0BPR"
     )
     r"""The unique identifier of the data source"""