+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -247,7 +247,7 @@ class ListDatasourcesData(BaseModel):
     metadata: ListDatasourcesMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY4MCVQNZXXV96J2V97BA90E"
+        "01KY4TS7NEYHJ4Y0PZH0AN2PV9"
     )
     r"""The unique identifier of the data source"""