+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -169,7 +169,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     metadata: RetrieveDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY4MCVQSCHVP9NVDQ6TV5Y7X"
+        "01KY4TS7NKS4HHE47GYT3GR3JC"
     )
     r"""The unique identifier of the data source"""