+++ packages/orq-rc/RELEASES.md
@@ -9639,3 +9639,13 @@ Based on:
 - [python v4.12.0-rc.52] packages/orq-rc
 ### Releases
 - [PyPI v4.12.0-rc.52] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.52 - packages/orq-rc
+
+## 2026-07-22 11:55:47
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.3 (2.918.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.12.0-rc.53] packages/orq-rc
+### Releases
+- [PyPI v4.12.0-rc.53] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.53 - packages/orq-rc